langchain1/dev/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=1.1.3
redis>=4.0.0
python-dotenv>=0.19.0
cryptography>=41.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiosmtplib>=1.1.6
langchain>=0.0.350
langchain-community>=0.0.3
langchain-experimental>=0.0.52
langchain-openai>=0.0.5
tavily-python
openai>=1.3.7
faiss-cpu>=1.7.4
wikipedia>=1.4.0
arxiv>=2.0.0
sentence-transformers>=2.2.0
pdfplumber>=0.10.0
backoff>=2.2.1
tenacity>=8.2.3
boto3>=1.34.0
transformers>=4.36.0
tiktoken>=0.5.1
unstructured>=0.10.0
# Document Loaders
python-docx
python-pptx
pypdf
beautifulsoup4
docx2txt
reportlab>=4.0.0  # Added for PDF generation